# flask/static/openapi.yaml
openapi: 3.0.3
info:
  title: Paws dataset API
  description: |-
    This API exposes the Paws dataset. Its data source:
    * The [Sonoma County Animal Shelter Intake and Outcome dataset](https://data.sonomacounty.ca.gov/Government/Base-stats/)
    
    The dataset contains around 18,000 pets (in the shelter), along with the arrival and outcome details related to the pet (intake type, intake date, intake condition, location, and similarly for the outcome, etc.).

    Brief summary on some information on the dataset: 

    intake_id:
    - 1:'STRAY'
    - 2:'CONFISCATE'
    - 3:'OWNER SURRENDER'
    - 4:'QUARANTINE'
    - 5:'ADOPTION RETURN'
    - 6:'TRANSFER' 
    - 7:'BORN HERE'

    sex_id:
    - 1:'Neutered'
    - 2:'Female'
    - 3:'Spayed'
    - 4:'Male'
    - 5:'Unknown'

    size_id:
    - 1:'LARGE'
    - 2:'SMALL'
    - 3:'MED'
    - 4:'TOY'
    - 5:'PUPPY'
    - 6:'KITTN'
    - 7:'X-LRG'

    type_id:
    - 1:'DOG'
    - 2:'CAT'
    - 3:'OTHER'
    
  The information available through this API can be accessed through the following paths:

paths:
  /pets/{animal_id}:
    get:
      summary: Get a pet by its ID
      description: Return the single pet associated with that ID
      parameters:
        - name: animal_id
          in: path
          description: ID of the pet
          example: A035813
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'          
        '404':
          description: Pet not found
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
        - api_auth: []
  /pets:
    get:
      summary: Get all pets
      description: Return all the information for all pets.
      parameters:
        - name: page
          in: query
          example: 0
          description: Index of the page (used for pagination)
          schema:
            type: integer
        - name: page_size
          in: query
          example: 100
          description: Number of pets per page (used for pagination)
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  movies:
                    type: array
                    items:
                      $ref: '#/components/schemas/Movie'
                  next_page:
                    type: integer
                    example: 1
                  last_page:
                    type: integer
                    example: 1000
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
        - api_auth: []
        
  securitySchemes:
    api_auth:
      type: http
      scheme: basic
      description: Please authenticate to use the API